黄建恒：将自定义的数据类型如：uint32之类的删除，修改DB访问流程

+++ MythCore/GameServer/inc/gameserver.h
@@ -41,7 +41,7 @@ public:
 
 public:
 	void		pushTask(EmTaskType eTaskType, CInternalMsg* pMsg);
-	void		pushDBTask(int nUid, uint8* pData, uint8 nDataLength);
+	void		pushDBTask(int nUid, byte* pData, int nDataLength);
 
 private:
 	CLog*					mDefaultLog;
